# Move unboxing helpers to managed code