--- a/src/coreclr/System.Private.CoreLib/src/System/Runtime/CompilerServices/BoxingHelpers.cs
+++ b/src/coreclr/System.Private.CoreLib/src/System/Runtime/CompilerServices/BoxingHelpers.cs
@@ -36,7 +36,7 @@ namespace System.Runtime.CompilerServices
                 return true;
             }
 
-            if (pMTa->HasTypeEquivalence && pMTb->HasTypeEquivalence)
+            if (!pMTa->HasTypeEquivalence || !pMTb->HasTypeEquivalence)
             {
                 return false;
             }