… src/coreclr/System.Private.CoreLib/src/System/Runtime/CompilerServices/RuntimeHelpers.CoreCLR.cs
… src/coreclr/inc/jithelpers.h
… src/coreclr/vm/JitQCallHelpers.h
… src/coreclr/vm/comutilnative.cpp
… src/coreclr/vm/comutilnative.h
… src/coreclr/vm/corelib.h
… src/coreclr/vm/ecalllist.h
… src/coreclr/vm/generics.cpp
… src/coreclr/vm/jithelpers.cpp
… src/coreclr/vm/jitinterface.h
… src/coreclr/vm/methodtable.h
… src/coreclr/vm/methodtable.inl
… src/coreclr/vm/methodtablebuilder.cpp
… src/coreclr/vm/object.cpp
… src/coreclr/vm/object.h
… src/coreclr/vm/object.inl
… src/coreclr/vm/qcallentrypoints.cpp