Move the unboxing helpers to managed code.

Behavior is basically identical except for the Unbox_Nullable paths, which required some investigation to find the fastest implementation. Notably, there interruptibility of managed code makes the copying/zeroing of values more difficult, but with the opportunity/requirement to specialize the codebase came a few micro-optimizations that are somewhat nice. Overall I don't expect anyone to notice the performance changes here, but since my earlier code was about 2X slower than the native implementation, I did feel the need to optimize until everything looked good.

Performance results:

| TestName | With PR | Without PR | % Speedup |
| --- | --- | --- | --- |
|                     TestJustAPrimitive| 1217.0047| 1221.203| 0.34%   |
|                         TestJustObject| 1212.6415| 1211.6437|-0.08%  |
|                     TestJust5Primitive| 1496.5304| 1522.4968|1.74%   |
|                        TestJust5Object| 1461.0507| 1488.3328|1.87%   |
|                    TestJust10Primitive| 1473.2814| 1493.5238|1.37%   |
|                       TestJust10Object| 3215.6339| 2854.6186|-11.23% |
|    TestJustAPrimitiveNullableWithValue| 2727.9085| 5182.2611|89.97%  |
|        TestJustObjectNullableWithValue| 3148.9484| 5672.2985|80.13%  |
|    TestJust5PrimitiveNullableWithValue| 5443.9232| 7795.6109|43.20%  |
|       TestJust5ObjectNullableWithValue| 6492.9071| 8095.1508|24.68%  |
|   TestJust10PrimitiveNullableWithValue| 6022.6274| 8723.572| 44.85%  |
|      TestJust10ObjectNullableWithValue| 7728.3239| 9671.1382|25.14%  |
|         TestJustAPrimitiveNullNullable| 1786.1337| 2230.0932|24.86%  |
|             TestJustObjectNullNullable| 1675.0683| 2326.0395|38.86%  |
|         TestJust5PrimitiveNullNullable| 2921.9497| 3298.4642|12.89%  |
|            TestJust5ObjectNullNullable| 3389.4043| 3615.3131|6.67%   |
|        TestJust10PrimitiveNullNullable| 3050.809|	 4054.9683|32.91%  |
|           TestJust10ObjectNullNullable| 4658.8316| 5335.0686|14.52%  |

Results are very positive, or within the margin of error in this test suite. These results were generated using a small benchmark which mostly targeted measuring the performance of the Unbox_Nullable helper, as it has the most complex and potentially slow code. Generally the impact on that helper is that the performance of the type system portion of the helper is faster, and the performance of code which actually copies the contents of a valuetype is a little better. This isn't quite a fair test of managed vs native performance though, as I took the opportunity to restructure some of the memory on `MethodTable` so that it could more easily be read in managed code, and that happened to make a fair bit of complex code become simpler and thus faster.

<details>
  <summary>Benchmark code (standalone console app)</summary>

```c#
using System.Diagnostics;
using System.Runtime.CompilerServices;

namespace UnboxingPerfTests
{
    internal class Program
    {
        public interface IStaticMethod<T>
        {
            public static abstract void Method(ref T param);
        }
        // All tests use GenericStruct<T, object> to force us into canonical code gen to always use the helpers
        struct GenericStruct<T, V> : IStaticMethod<GenericStruct<T, V>> where T: IStaticMethod<T>
        {
            public T Value;

            [MethodImpl(MethodImplOptions.NoInlining)]
            public static void Method(ref GenericStruct<T, V> value)
            {
                T.Method(ref value.Value);
            }
        }

        struct JustAPrimitive : IStaticMethod<JustAPrimitive>
        {
            public int Value;
            public static void Method(ref JustAPrimitive param)
            {
            }
        }

        struct JustObject : IStaticMethod<JustObject>
        {
            public object Value;
            public static void Method(ref JustObject param)
            {
            }
        }

        struct Just10Primitive : IStaticMethod<Just10Primitive>
        {
            public int Value;
            public int Value2;
            public int Value3;
            public int Value4;
            public int Value5;
            public int Value6;
            public int Value7;
            public int Value8;
            public int Value9;
            public int Value10;
            public static void Method(ref Just10Primitive param)
            {
            }
        }

        struct Just10Object : IStaticMethod<Just10Object>
        {
            public object Value;
            public object Value2;
            public object Value3;
            public object Value4;
            public object Value5;
            public object Value6;
            public object Value7;
            public object Value8;
            public object Value9;
            public object Value10;

            public static void Method(ref Just10Object param)
            {
            }
        }

        struct Just5Primitive : IStaticMethod<Just5Primitive>
        {
            public int Value;
            public int Value2;
            public int Value3;
            public int Value4;
            public int Value5;
            public static void Method(ref Just5Primitive param)
            {
            }
        }

        struct Just5Object : IStaticMethod<Just5Object>
        {
            public object Value;
            public object Value2;
            public object Value3;
            public object Value4;
            public object Value5;

            public static void Method(ref Just5Object param)
            {
            }
        }


        public static void Unbox_any_Test<T>(object[] o) where T:IStaticMethod<T>
        {
            for (int i = 0; i < 1_000_000; i++)
            {
                T local = (T)o[i%3];
                T.Method(ref local);
            }
        }

        [MethodImpl(MethodImplOptions.NoInlining)]
        public static int IsInst_Test<T>(object?[] o)
        {
            int retVal = 0;
            for (int i = 0; i < 1_000_000; i++)
            {
                if (o[i % 3] is T)
                    retVal++;
            }
            return retVal;
        }

        public static void Unbox_any_TestNullable<T>(object?[] o) where T : struct, IStaticMethod<T>
        {
            for (int i = 0; i < 1_000_000; i++)
            {
                T? local = (T?)o[i % 3];
                if (local.HasValue)
                {
                    T localCopy = local.Value;
                    T.Method(ref localCopy);
                }
            }
        }


        [MethodImpl(MethodImplOptions.AggressiveOptimization)]
        public static void Test(int outerIterationCount, Action a, string testName, Action<string> outputFunc)
        {
            Stopwatch stopwatch = Stopwatch.StartNew();
            for (int i = 0; i < outerIterationCount; i++)
            {
                a();
            }
            stopwatch.Stop();
            outputFunc($"| {testName,38}| {stopwatch.Elapsed.TotalMilliseconds.ToString()}|");
        }

        static void TestJustAPrimitive()
        {
            object[] arr = { new GenericStruct<JustAPrimitive, object>(), new GenericStruct<JustAPrimitive, object>(), new GenericStruct<JustAPrimitive, object>() };
            Unbox_any_Test<GenericStruct<JustAPrimitive, object>>(arr);
        }

        static void TestJustObject()
        {
            object[] arr = { new GenericStruct<JustObject, object>(), new GenericStruct<JustObject, object>(), new GenericStruct<JustObject, object>() };
            Unbox_any_Test<GenericStruct<JustObject, object>>(arr);
        }

        static void TestJust10Primitive()
        {
            object[] arr = { new GenericStruct<Just10Primitive, object>(), new GenericStruct<Just10Primitive, object>(), new GenericStruct<Just10Primitive, object>() };
            Unbox_any_Test<GenericStruct<Just10Primitive, object>>(arr);
        }

        static void TestJust10Object()
        {
            object[] arr = { new GenericStruct<Just10Object, object>(), new GenericStruct<Just10Object, object>(), new GenericStruct<Just10Object, object>() };
            Unbox_any_Test<GenericStruct<Just10Object, object>>(arr);
        }

        static void TestJust5Primitive()
        {
            object[] arr = { new GenericStruct<Just5Primitive, object>(), new GenericStruct<Just5Primitive, object>(), new GenericStruct<Just5Primitive, object>() };
            Unbox_any_Test<GenericStruct<Just5Primitive, object>>(arr);
        }

        static void TestJust5Object()
        {
            object[] arr = { new GenericStruct<Just5Object, object>(), new GenericStruct<Just5Object, object>(), new GenericStruct<Just5Object, object>() };
            Unbox_any_Test<GenericStruct<Just5Object, object>>(arr);
        }


        static void TestJustAPrimitiveNullableWithValue()
        {
            object[] arr = { new GenericStruct<JustAPrimitive, object>(), new GenericStruct<JustAPrimitive, object>(), new GenericStruct<JustAPrimitive, object>() };
            Unbox_any_TestNullable<GenericStruct<JustAPrimitive, object>>(arr);
        }

        static void TestJustObjectNullableWithValue()
        {
            object[] arr = { new GenericStruct<JustObject, object>(), new GenericStruct<JustObject, object>(), new GenericStruct<JustObject, object>() };
            Unbox_any_TestNullable<GenericStruct<JustObject, object>>(arr);
        }

        static void TestJust10PrimitiveNullableWithValue()
        {
            object[] arr = { new GenericStruct<Just10Primitive, object>(), new GenericStruct<Just10Primitive, object>(), new GenericStruct<Just10Primitive, object>() };
            Unbox_any_TestNullable<GenericStruct<Just10Primitive, object>>(arr);
        }

        static void TestJust10ObjectNullableWithValue()
        {
            object[] arr = { new GenericStruct<Just10Object, object>(), new GenericStruct<Just10Object, object>(), new GenericStruct<Just10Object, object>() };
            Unbox_any_TestNullable<GenericStruct<Just10Object, object>>(arr);
        }

        static void TestJust5PrimitiveNullableWithValue()
        {
            object[] arr = { new GenericStruct<Just5Primitive, object>(), new GenericStruct<Just5Primitive, object>(), new GenericStruct<Just5Primitive, object>() };
            Unbox_any_TestNullable<GenericStruct<Just5Primitive, object>>(arr);
        }

        static void TestJust5ObjectNullableWithValue()
        {
            object[] arr = { new GenericStruct<Just5Object, object>(), new GenericStruct<Just5Object, object>(), new GenericStruct<Just5Object, object>() };
            Unbox_any_TestNullable<GenericStruct<Just5Object, object>>(arr);
        }


        static void TestJustAPrimitiveNullNullable()
        {
            object?[] arr = { null, null, null };
            Unbox_any_TestNullable<GenericStruct<JustAPrimitive, object>>(arr);
        }

        static void TestJustObjectNullNullable()
        {
            object?[] arr = { null, null, null };
            Unbox_any_TestNullable<GenericStruct<JustObject, object>>(arr);
        }

        static void TestJust10PrimitiveNullNullable()
        {
            object?[] arr = { null, null, null };
            Unbox_any_TestNullable<GenericStruct<Just10Primitive, object>>(arr);
        }

        static void TestJust10ObjectNullNullable()
        {
            object?[] arr = { null, null, null };
            Unbox_any_TestNullable<GenericStruct<Just10Object, object>>(arr);
        }

        static void TestJust5PrimitiveNullNullable()
        {
            object?[] arr = { null, null, null };
            Unbox_any_TestNullable<GenericStruct<Just5Primitive, object>>(arr);
        }

        static void TestJust5ObjectNullNullable()
        {
            object?[] arr = { null, null, null };
            Unbox_any_TestNullable<GenericStruct<Just5Object, object>>(arr);
        }

        static void IsInstTestJustAPrimitive()
        {
            object[] arr = { new GenericStruct<JustAPrimitive, object>(), new GenericStruct<JustAPrimitive, object>(), new GenericStruct<JustAPrimitive, object>() };
            IsInst_Test<GenericStruct<JustAPrimitive, object>>(arr);
        }

        static void IsInstTestJustObject()
        {
            object[] arr = { new GenericStruct<JustObject, object>(), new GenericStruct<JustObject, object>(), new GenericStruct<JustObject, object>() };
            IsInst_Test<GenericStruct<JustObject, object>>(arr);
        }

        static void IsInstTestJust10Primitive()
        {
            object[] arr = { new GenericStruct<Just10Primitive, object>(), new GenericStruct<Just10Primitive, object>(), new GenericStruct<Just10Primitive, object>() };
            IsInst_Test<GenericStruct<Just10Primitive, object>>(arr);
        }

        static void IsInstTestJust10Object()
        {
            object[] arr = { new GenericStruct<Just10Object, object>(), new GenericStruct<Just10Object, object>(), new GenericStruct<Just10Object, object>() };
            IsInst_Test<GenericStruct<Just10Object, object>>(arr);
        }

        static void IsInstTestJust5Primitive()
        {
            object[] arr = { new GenericStruct<Just5Primitive, object>(), new GenericStruct<Just5Primitive, object>(), new GenericStruct<Just5Primitive, object>() };
            IsInst_Test<GenericStruct<Just5Primitive, object>>(arr);
        }

        static void IsInstTestJust5Object()
        {
            object[] arr = { new GenericStruct<Just5Object, object>(), new GenericStruct<Just5Object, object>(), new GenericStruct<Just5Object, object>() };
            IsInst_Test<GenericStruct<Just5Object, object>>(arr);
        }


        static void IsInstTestJustAPrimitiveNullableWithValue()
        {
            object[] arr = { new GenericStruct<JustAPrimitive, object>(), new GenericStruct<JustAPrimitive, object>(), new GenericStruct<JustAPrimitive, object>() };
            IsInst_Test<GenericStruct<JustAPrimitive, object>>(arr);
        }

        static void IsInstTestJustObjectNullableWithValue()
        {
            object[] arr = { new GenericStruct<JustObject, object>(), new GenericStruct<JustObject, object>(), new GenericStruct<JustObject, object>() };
            IsInst_Test<GenericStruct<JustObject, object>>(arr);
        }

        static void IsInstTestJust10PrimitiveNullableWithValue()
        {
            object[] arr = { new GenericStruct<Just10Primitive, object>(), new GenericStruct<Just10Primitive, object>(), new GenericStruct<Just10Primitive, object>() };
            IsInst_Test<GenericStruct<Just10Primitive, object>?>(arr);
        }

        static void IsInstTestJust10ObjectNullableWithValue()
        {
            object[] arr = { new GenericStruct<Just10Object, object>(), new GenericStruct<Just10Object, object>(), new GenericStruct<Just10Object, object>() };
            IsInst_Test<GenericStruct<Just10Object, object>?>(arr);
        }

        static void IsInstTestJust5PrimitiveNullableWithValue()
        {
            object[] arr = { new GenericStruct<Just5Primitive, object>(), new GenericStruct<Just5Primitive, object>(), new GenericStruct<Just5Primitive, object>() };
            IsInst_Test<GenericStruct<Just5Primitive, object>?>(arr);
        }

        static void IsInstTestJust5ObjectNullableWithValue()
        {
            object[] arr = { new GenericStruct<Just5Object, object>(), new GenericStruct<Just5Object, object>(), new GenericStruct<Just5Object, object>() };
            IsInst_Test<GenericStruct<Just5Object, object>?>(arr);
        }


        static void IsInstTestJustAPrimitiveNullNullable()
        {
            object?[] arr = { null, null, null };
            IsInst_Test<GenericStruct<JustAPrimitive, object>?>(arr);
        }

        static void IsInstTestJustObjectNullNullable()
        {
            object?[] arr = { null, null, null };
            IsInst_Test<GenericStruct<JustObject, object>?>(arr);
        }

        static void IsInstTestJust10PrimitiveNullNullable()
        {
            object?[] arr = { null, null, null };
            IsInst_Test<GenericStruct<Just10Primitive, object>?>(arr);
        }

        static void IsInstTestJust10ObjectNullNullable()
        {
            object?[] arr = { null, null, null };
            IsInst_Test<GenericStruct<Just10Object, object>?>(arr);
        }

        static void IsInstTestJust5PrimitiveNullNullable()
        {
            object?[] arr = { null, null, null };
            IsInst_Test<GenericStruct<Just5Primitive, object>?>(arr);
        }

        static void IsInstTestJust5ObjectNullNullable()
        {
            object?[] arr = { null, null, null };
            IsInst_Test<GenericStruct<Just5Object, object>?>(arr);
        }

        enum Foo : int
        {
            Value = 3
        }

        enum Bar : int
        {

        }

        enum Other : int
        {
            Value = 4,
        }

        public static object[] TheArrayForEnumTests = { (int)4, Foo.Value, Other.Value };
        static void IsInstTestEnumVariance()
        {
            int count = 0;
            object[] arr = TheArrayForEnumTests;

            for (int i = 0; i < 1_000_000; i++)
            {
                Bar local = (Bar)arr[i % 3];
                count += (int)local;
            }

            s_x = count;
        }
        public static int s_x;
        static void IsInstTestEnumVarianceNullable()
        {
            int count = 0;
            object?[] arr = TheArrayForEnumTests;

            for (int i = 0; i < 1_000_000; i++)
            {
                Bar? local = (Bar?)arr[i % 3];
                if (local.HasValue)
                {
                    count += (int)local.Value;
                }
            }
            s_x = count;
        }

        static void AllTests(int iterationCount, Action<string> outputFunc)
        {
/*            Test(iterationCount, IsInstTestJustAPrimitive, nameof(IsInstTestJustAPrimitive), outputFunc);
            Test(iterationCount, IsInstTestJustObject, nameof(IsInstTestJustObject), outputFunc);
            Test(iterationCount, IsInstTestJust5Primitive, nameof(IsInstTestJust5Primitive), outputFunc);
            Test(iterationCount, IsInstTestJust5Object, nameof(IsInstTestJust5Object), outputFunc);
            Test(iterationCount, IsInstTestJust10Primitive, nameof(IsInstTestJust10Primitive), outputFunc);
            Test(iterationCount, IsInstTestJust10Object, nameof(IsInstTestJust10Object), outputFunc);

            Test(iterationCount, IsInstTestJustAPrimitiveNullableWithValue, nameof(IsInstTestJustAPrimitiveNullableWithValue), outputFunc);
            Test(iterationCount, IsInstTestJustObjectNullableWithValue, nameof(IsInstTestJustObjectNullableWithValue), outputFunc);*/
/*            Test(iterationCount, IsInstTestJust5PrimitiveNullableWithValue, nameof(IsInstTestJust5PrimitiveNullableWithValue), outputFunc);
            Test(iterationCount, IsInstTestJust5ObjectNullableWithValue, nameof(IsInstTestJust5ObjectNullableWithValue), outputFunc);
            Test(iterationCount, IsInstTestJust10PrimitiveNullableWithValue, nameof(IsInstTestJust10PrimitiveNullableWithValue), outputFunc);
            Test(iterationCount, IsInstTestJust10ObjectNullableWithValue, nameof(IsInstTestJust10ObjectNullableWithValue), outputFunc);

            Test(iterationCount, IsInstTestJustAPrimitiveNullNullable, nameof(IsInstTestJustAPrimitiveNullNullable), outputFunc);
…